backend/requirements.txt
Flask==1.1.1
logzero==1.3.1
requests==2.14.2
pymongo==3.10.1
python-etcd==0.4.5